Store nonce used when generating an AEAD when saving an AEAD to file.
If a nonce different from the public_id is used when generating an AEAD,
we must keep track of the nonce separately. This commit changes the
file format used, but should be backwards compatible by using the
public_id as nonce on load() if the old file format is detected.